Refactor GPU mappers and standardize enum naming.
Replaced deprecated mapper methods with updated implementations to improve clarity and consistency. Standardized enum naming conventions and fixed typos to align with coding standards. Removed unsupported vertex formats and unused methods for improved maintainability.

+++ wgpu4k/src/commonMain/kotlin/Types.kt
@@ -8,7 +8,7 @@ typealias GPUColorWriteFlags = Int
 
 typealias GPUDepthBias = Int
 
-typealias GPUFlagsConstant = ULong
+typealias GPUFlagsConstant = UInt
 
 typealias GPUIndex32 = UInt
