+++ wgpu4k/src/jsMain/kotlin/io.ygdrasil.wgpu/Buffer.js.kt
@@ -10,7 +10,7 @@ actual class Buffer(internal val handler: GPUBuffer) : AutoCloseable {
     actual val size: GPUSize64
         get() = handler.size
     actual val usage: Set<BufferUsage>
-        get() = BufferUsage.entries.filter { it.value.toULong() and handler.usage != 0uL }.toSet()
+        get() = BufferUsage.entries.filter { it.value.toUInt() and handler.usage != 0u }.toSet()
     actual val mapState: BufferMapState
         get() = BufferMapState.of(handler.mapState) ?: error("fail to get MapState")
